Add color

+++ xsh/cmd/guardian/main.go
@@ -2,27 +2,27 @@ package main
 
 import (
 	"bytes"
-	"io/ioutil"
+	"github.com/chzyer/readline"
 	"io"
+	"io/ioutil"
 	"log"
 	"os"
 	"os/exec"
-	"github.com/chzyer/readline"
 	"time"
 
 	"github.com/nsf/termbox-go"
 )
-
 
 var SubProcHandle *exec.Cmd
 var InterActiveSub bool
 var StopCopy chan bool
 
 type subprocpipes struct {
-	Stdin io.WriteCloser
+	Stdin  io.WriteCloser
 	Stdout io.ReadCloser
 	Stderr io.ReadCloser
 }
+
 //Copy from reader to writer
 func CopyFilter(reader io.Reader, writer io.Writer) {
 	log.Printf("Started copyfilter")
@@ -32,9 +32,9 @@ func CopyFilter(reader io.Reader, writer io.Writer) {
 		if err != nil {
 			break
 		}
-		if bytes.Contains(buf[:n], []byte{3} ) || 
-		bytes.Contains(buf[:n], []byte{4} ) || 
-		bytes.Contains(buf[:n], []byte("z") ) {
+		if bytes.Contains(buf[:n], []byte{3}) ||
+			bytes.Contains(buf[:n], []byte{4}) ||
+			bytes.Contains(buf[:n], []byte("z")) {
 			log.Println("Ctrl-C detected, exiting")
 			SubProcHandle.Process.Kill()
 			termbox.Clear(termbox.ColorDefault, termbox.ColorDefault)
@@ -61,35 +61,38 @@ func QuickCommandInteractivePrep(strs []string) (*subprocpipes, *exec.Cmd) {
 	return &sub, cmd
 }
 
-func main () {
+func main() {
 	log.SetOutput(ioutil.Discard)
 	//termbox.Init()
 	command := os.Args[1:]
-	_,cmd := QuickCommandInteractivePrep(command)
+	_, cmd := QuickCommandInteractivePrep(command)
 	SubProcHandle = cmd
-	
+
 	log.Printf("Guardian: starting %+v\n", command)
 	l, _ := readline.New("")
 	l.Terminal.EnterRawMode()
-	cmd.Start()
+	err := cmd.Start()
 	/*
-	for cmd.ProcessState==nil {
-		time.Sleep(time.Second)
-	}
-	for ! cmd.ProcessState.Exited() {
-		time.Sleep(time.Second)
-	}
+		for cmd.ProcessState==nil {
+			time.Sleep(time.Second)
+		}
+		for ! cmd.ProcessState.Exited() {
+			time.Sleep(time.Second)
+		}
 	*/
-	
-	cmd.Process.Wait()
+	if err == nil {
+		cmd.Process.Wait()
 
-	log.Println("Subprocess exited, exiting guardian now")
-	time.Sleep(time.Millisecond)
-	//termbox.Clear(termbox.ColorDefault, termbox.ColorDefault)
-	//termbox.Close()
-	//l.Terminal.Close()
+		log.Println("Subprocess exited, exiting guardian now")
+		time.Sleep(time.Millisecond)
+		//termbox.Clear(termbox.ColorDefault, termbox.ColorDefault)
+		//termbox.Close()
+		//l.Terminal.Close()
 
-	l.Terminal.ExitRawMode()
-	os.Exit(0)
-	
+		l.Terminal.ExitRawMode()
+		os.Exit(0)
+	} else {
+		log.Println("Error: ", err)
+	}
+
 }
